requirements.txt
sentence-transformers==2.2.2
sentence-transformers
langchain
flask
pypdf
python-dotenv
pinecone[grpc]
langchain-pinecone
langchain_community
langchain_experimental
langchain-google-genai
google-generativeai

-e .